+++ datatig/models/field_date.py
@@ -1,3 +1,7 @@
+import datetime
+
+import dateparser
+
 from datatig.jsondeepreaderwriter import JSONDeepReaderWriter
 from datatig.models.field import FieldConfigModel, FieldValueModel
 
@@ -28,19 +32,51 @@ class FieldDateConfigModel(FieldConfigModel):
                 "name": "field_" + self.get_id(),
                 "title": self.get_title(),
                 "type": "date",
-            }
+            },
+            {
+                "name": "field_" + self.get_id() + "___timestamp",
+                "title": self.get_title() + " (Timestamp)",
+                "type": "integer",
+            },
         ]
 
 
 class FieldDateValueModel(FieldValueModel):
     def set_value(self, value):
-        self._value = value
+        self._value = None
+        if isinstance(value, str):
+            self._value = dateparser.parse(value, settings={"TIMEZONE": "UTC"})
+            if self._value:
+                self._value = self._value.date()
+        elif isinstance(value, datetime.date):
+            self._value = value
+        elif isinstance(value, datetime.datetime):
+            self._value = value.date()
 
     def get_value(self):
-        return self._value
+        if self._value:
+            return self._value.isoformat()
+        else:
+            return None
+
+    def get_value_timestamp(self):
+        if self._value:
+            dt = datetime.datetime(
+                self._value.year,
+                self._value.month,
+                self._value.day,
+                12,
+                0,
+                0,
+                0,
+                datetime.timezone.utc,
+            )
+            return dt.timestamp()
+        else:
+            return None
 
     def get_frictionless_csv_data_values(self):
-        return [self._value]
+        return [self.get_value(), self.get_value_timestamp()]
 
     def different_to(self, other_field_value):
         return self._value != other_field_value._value